+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+* Added Save to stream method
+* Added other Load methods
+
 ## [0.5.0] - 2025-02-20
 
 ### Added